+++ packages/transports/new-relic/src/index.ts
@@ -1,1 +1,3 @@
-export * from "./NewRelicTransport.js";
+export { HttpTransportError, LogSizeError, RateLimitError } from "@loglayer/transport-http";
+export type { NewRelicTransportConfig } from "./NewRelicTransport.js";
+export { NewRelicTransport, ValidationError } from "./NewRelicTransport.js";